backend/backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
dotenv==0.9.9
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc0
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-genai==1.12.1
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.72.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9.1
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
websockets==15.0.1
